h264: build frame-mode ref lists from combined field pairs for mixed PAFF streams
build_ref_list_l0 (frame-mode P/B slice RefPicList0) iterated raw DPB
entries directly. For a stream that mixes PAFF field pictures with frame
pictures (frame_mbs_only_flag == 0 but not every picture sets
field_pic_flag, e.g. ITU Sharp_MP_PAFF_1r2 / CVPA1_TOSHIBA_B /
CAPA1_TOSHIBA_B), the DPB can hold raw half-height field entries when a
later frame-coded P slice needs a reference. Handing those straight to
frame-mode motion compensation as if they were full-height frames
corrupted ~90% of the picture (wrong stride/sample positions).

Add combine_field_pairs_into_frames: pairs up complementary field DPB
entries (same frame_num, opposite bottom_field_flag, matching short/
long-term status) and interleaves each pair into one full-height frame
DpbEntry via the existing H264Decoder::interleave_fields helper (the
same routine the output field-pairing path already uses), per
§8.2.4.2.1's "reference frames and complementary reference field pairs"
construction rule. An unpaired field is dropped from the frame-mode list
but remains usable by later field pictures via build_field_ref_list_l0,
which still reads the DPB directly.

Verified the constructed reference is byte-exact against the ITU
reference YUV in isolation. ITU conformance suite: still 32/32
hard-checked BitExact clips, 0 regressions; Sharp_MP_PAFF_1r2/
CAPA1_TOSHIBA_B/CVPA1_TOSHIBA_B all measurably improved (more exact
frames, fewer diff bytes) though none newly bit-exact — see todo-h264.md
session addendum for the remaining gap.

Also fixes two clippy -D warnings violations blocking `just check`
(i32::abs() as u64 -> unsigned_abs in reconstruct.rs; redundant as i32
cast in sps.rs's POC-type-1 parsing) and a benches/decode_throughput.rs
SeqParameterSet literal that hadn't been updated for the new POC-type-1
fields, plus a pending fmt-only cleanup in dbg_field_triage.rs.

+++ tpt-kinetix-h264/src/sps.rs
@@ -122,9 +122,10 @@ impl SeqParameterSet {
             offset_for_non_ref_pic = r.read_se().context("offset_for_non_ref_pic")?;
             offset_for_top_to_bottom_field =
                 r.read_se().context("offset_for_top_to_bottom_field")?;
-            num_ref_frames_in_pic_order_cnt_cycle = r.read_ue().context("num_ref_frames_in_poc_cycle")?;
+            num_ref_frames_in_pic_order_cnt_cycle =
+                r.read_ue().context("num_ref_frames_in_poc_cycle")?;
             for _ in 0..num_ref_frames_in_pic_order_cnt_cycle {
-                offset_for_ref_frame.push(r.read_se().context("offset_for_ref_frame")? as i32);
+                offset_for_ref_frame.push(r.read_se().context("offset_for_ref_frame")?);
             }
         }
